add relattionsip in all tables

+++ src/services/competition_crud.py
@@ -1,4 +1,4 @@
-from sqlalchemy.orm import Session
+from sqlalchemy.orm import Session, joinedload
 from src.models.competition import Competition
 from src.models.user import User
 from src.schemas.competition import CompetitionCreateSchema, CompetitionUpdateSchema
@@ -9,7 +9,7 @@ def get_all_competition(db:Session, skip:int = 0, limit:int = 100):
 
 
 def create_competition(db:Session, competition:CompetitionCreateSchema):
-    new_competition = Competition(user_id=competition.user_id, name=competition.name)
+    new_competition = Competition(name=competition.name)
     db.add(new_competition)
     db.commit()
     db.refresh(new_competition)
@@ -17,24 +17,32 @@ def create_competition(db:Session, competition:CompetitionCreateSchema):
 
 
 def get_competition_by_id(db:Session, competition_id:int):
+    # data = db.query(Competition).options(joinedload(User.id)).first()
+    # print("========",data)
     return db.query(Competition).filter(Competition.id == competition_id).first()
     
 
 def update_competiton(db:Session, competition:CompetitionUpdateSchema, id:int):
     comp = get_competition_by_id(db,id)
-    comp.id = comp.id
-    comp.name = competition.name
-    comp.user_id = competition.user_id
-    db.add(comp)
-    db.commit()
-    db.refresh(comp)
-    return comp
+    if comp:
+        comp.id = comp.id
+        comp.name = competition.name
+        db.add(comp)
+        db.commit()
+        db.refresh(comp)
+        return {"updated_competition":comp, "message":"Competition is updated"}
+    else:
+        return {"message":"Competition not found"}
 
 
 def delete_competition(db:Session, competition_id:int):
     com = get_competition_by_id(db,competition_id)
-    db.delete(com)
-    db.commit()
+    if com:
+        db.delete(com)
+        db.commit()
+        return "competition deleted"
+    else:
+        return "competition not found"
